OMAPI/TMAPI: Set MSS on es10x commands to 0xFF for channels
https://android.googlesource.com/platform/frameworks/opt/telephony/+/master/src/java/com/android/internal/telephony/uicc/euicc/apdu/RequestBuilder.java#31

+++ libs/lpac-jni/src/main/java/net/typeblog/lpac_jni/impl/LocalProfileAssistantImpl.kt
@@ -90,8 +90,8 @@ class LocalProfileAssistantImpl(
         httpInterface.usePublicKeyIds(pkids.toTypedArray())
     }
 
-    override fun setEs10xMss(mss: Byte) {
-        LpacJni.euiccSetMss(contextHandle, mss)
+    override fun setEs10xMss(mss: UByte) {
+        LpacJni.euiccSetMss(contextHandle, mss.toInt())
     }
 
     override val valid: Boolean
